Move res/enkf/config tests to match src/
