+++ tests/libres_tests/res/enkf/config/test_gen_data_config.py
@@ -1,0 +1,5 @@
+from ert._c_wrappers.enkf.config import GenDataConfig
+
+
+def test_create():
+    GenDataConfig("KEY")
