+++ tests/libres_tests/res/enkf/config/test_gen_kw_config.py
